Split out the Tweepy Wrapping into another file / class (this would make it easier to re-use the TweepyWrapper in another project later. Also split out the command line interface into a seperate file as this is not needed when using the class in another application.
