+++ cheer_lights_twitter_api/__init__.py
@@ -1,0 +1,6 @@
+"""
+Package to generate tweets to the CheerLights, this is designed to be used in an IoT application so that an embedded
+device can send a tweet
+"""
+from .cheer_lights_twitter_api import CheerLightTwitterAPI
+from .cheer_lights_twitter_api import CheerLightColours
